Template variables and first Text-Template
A few additions for variables. A loop now generates sentences and paragraph variables for the parser with array presets. This allows to define h1-h10 and paragraph 1-10 within the html templates. The File Alice now delivers a list of main-words filler-words and complicated words. These words are used by Rabbithole.php to generate words, unicode, sentences and paragraphs.